requirements.txt
appdirs==1.4.3
argon2-cffi==19.2.0
asgiref==3.2.7
astroid==2.3.3
awsebcli==3.17.1
botocore==1.14.17
cement==2.8.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.3.9
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.2.1
django-extensions==2.2.9
djangorestframework==3.11.0
docutils==0.15.2
filelock==3.0.12
future==0.16.0
gunicorn==20.0.4
idna==2.7
isort==4.3.21
jedi==0.16.0
jmespath==0.9.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.6.2
pathspec==0.5.9
prompt-toolkit==3.0.5
psycopg2==2.8.4
ptpython==3.0.1
pycparser==2.20
Pygments==2.6.1
pylint==2.4.4
pypiwin32==223
python-dateutil==2.8.0
pytz==2019.3
pywin32==227
PyYAML==5.2
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.1
stripe==2.44.0
termcolor==1.1.0
urllib3==1.24.3
virtualenv==20.0.3
wcwidth==0.1.9
whitenoise==5.0.1
wrapt==1.11.2
